finish radeon and USB card detect

+++ include/radeonfb.h
@@ -535,9 +535,9 @@ extern void radeonfb_pm_exit(struct radeonfb_info *rinfo);
 
 /* Monitor probe functions */
 extern void radeon_probe_screens(struct radeonfb_info *rinfo,
-                                 const char *monitor_layout, int ignore_edid);
+                                 const char *monitor_layout, int32_t ignore_edid);
 extern void radeon_check_modes(struct radeonfb_info *rinfo, struct mode_option *resolution);
-extern int radeon_match_mode(struct radeonfb_info *rinfo,
+extern int32_t radeon_match_mode(struct radeonfb_info *rinfo,
                              struct fb_var_screeninfo *dest,
                              const struct fb_var_screeninfo *src);
 
@@ -575,7 +575,7 @@ extern void RADEONVIP_reset(struct radeonfb_info *rinfo);
 
 /* Accel functions */
 
-extern void radeon_wait_for_fifo_function(struct radeonfb_info *rinfo, int entries);
+extern void radeon_wait_for_fifo_function(struct radeonfb_info *rinfo, int32_t entries);
 extern void radeon_engine_flush(struct radeonfb_info *rinfo);
 extern void radeon_engine_reset(struct radeonfb_info *rinfo);
 extern void radeon_engine_restore(struct radeonfb_info *rinfo);
@@ -592,7 +592,7 @@ extern void radeon_wait_for_idle_mmio(struct radeonfb_info *rinfo);
         rinfo->fifo_slots -= entries;							\
     } while (0)
 
-static inline int radeonfb_sync(struct fb_info *info)
+static inline int32_t radeonfb_sync(struct fb_info *info)
 {
     struct radeonfb_info *rinfo = info->par;
     radeon_engine_idle();
@@ -600,44 +600,44 @@ static inline int radeonfb_sync(struct fb_info *info)
 }
 
 extern void radeon_restore_accel_state_mmio(struct fb_info *info);
-extern void radeon_setup_for_solid_fill(struct fb_info *info, int color, int rop, unsigned int planemask);
-extern void radeon_subsequent_solid_fill_rect_mmio(struct fb_info *info, int x, int y, int w, int h);
-extern void radeon_setup_for_solid_line_mmio(struct fb_info *info, int color, int rop, unsigned int planemask);
-extern void radeon_subsequent_solid_hor_vert_line_mmio(struct fb_info *info, int x, int y, int len, int dir);
-extern void radeon_subsequent_solid_two_point_line_mmio(struct fb_info *info, int xa, int ya, int xb,
-                                                        int yb, int flags);
-extern void radeon_setup_for_dashed_line_mmio(struct fb_info *info, int fg, int bg,
-                                              int rop, unsigned int planemask, int length, unsigned char *pattern);
+extern void radeon_setup_for_solid_fill(struct fb_info *info, int32_t color, int32_t rop, uint32_t planemask);
+extern void radeon_subsequent_solid_fill_rect_mmio(struct fb_info *info, int32_t x, int32_t y, int32_t w, int32_t h);
+extern void radeon_setup_for_solid_line_mmio(struct fb_info *info, int32_t color, int32_t rop, uint32_t planemask);
+extern void radeon_subsequent_solid_hor_vert_line_mmio(struct fb_info *info, int32_t x, int32_t y, int32_t len, int32_t dir);
+extern void radeon_subsequent_solid_two_point_line_mmio(struct fb_info *info, int32_t xa, int32_t ya, int32_t xb,
+                                                        int32_t yb, int32_t flags);
+extern void radeon_setup_for_dashed_line_mmio(struct fb_info *info, int32_t fg, int32_t bg,
+                                              int32_t rop, uint32_t planemask, int32_t length, unsigned char *pattern);
 extern void radeon_subsequent_dashed_two_point_line_mmio(struct fb_info *info,
-                                                         int xa, int ya, int xb, int yb, int flags, int phase);
+                                                         int32_t xa, int32_t ya, int32_t xb, int32_t yb, int32_t flags, int32_t phase);
 extern void radeon_setup_for_screen_to_screen_copy_mmio(struct fb_info *info,
-                                                        int xdir, int ydir, int rop, unsigned int planemask, int trans_color);
+                                                        int32_t xdir, int32_t ydir, int32_t rop, uint32_t planemask, int32_t trans_color);
 extern void radeon_subsequent_screen_to_screen_copy_mmio(struct fb_info *info,
-                                                         int xa, int ya, int xb, int yb, int w, int h);
+                                                         int32_t xa, int32_t ya, int32_t xb, int32_t yb, int32_t w, int32_t h);
 extern void radeon_screen_to_screen_copy_mmio(struct fb_info *info,
-                                              int xa, int ya, int xb, int yb, int w, int h, int rop);
+                                              int32_t xa, int32_t ya, int32_t xb, int32_t yb, int32_t w, int32_t h, int32_t rop);
 extern void radeon_setup_for_mono_8x8_pattern_fill_mmio(struct fb_info *info,
-                                                        int patternx, int patterny, int fg, int bg, int rop, unsigned int planemask);
+                                                        int32_t patternx, int32_t patterny, int32_t fg, int32_t bg, int32_t rop, uint32_t planemask);
 extern void radeon_subsequent_mono_8x8_pattern_fill_rect_mmio(struct fb_info *info,
-                                                              int patternx, int patterny, int x, int y, int w, int h);
+                                                              int32_t patternx, int32_t patterny, int32_t x, int32_t y, int32_t w, int32_t h);
 extern void radeon_setup_for_scanline_cpu_to_screen_color_expand_fill_mmio(struct fb_info *info,
-                                                                           int fg, int bg, int rop, unsigned int planemask);
+                                                                           int32_t fg, int32_t bg, int32_t rop, uint32_t planemask);
 extern void radeon_subsequent_scanline_cpu_to_screen_color_expand_fill_mmio(struct fb_info *info,
-                                                                            int x, int y, int w, int h, int skipleft);
-extern void radeon_subsequent_scanline_mmio(struct fb_info *info, unsigned long *buf);
+                                                                            int32_t x, int32_t y, int32_t w, int32_t h, int32_t skipleft);
+extern void radeon_subsequent_scanline_mmio(struct fb_info *info, uint32_t *buf);
 extern void radeon_setup_for_scanline_image_write_mmio(struct fb_info *info,
-                                                       int rop, unsigned int planemask, int trans_color, int bpp);
+                                                       int32_t rop, uint32_t planemask, int32_t trans_color, int32_t bpp);
 extern void radeon_subsequent_scanline_image_write_rect_mmio(struct fb_info *info,
-                                                             int x, int y, int w, int h, int skipleft);
+                                                             int32_t x, int32_t y, int32_t w, int32_t h, int32_t skipleft);
 extern void radeon_set_clipping_rectangle_mmio(struct fb_info *info,
-                                               int xa, int ya, int xb, int yb);
+                                               int32_t xa, int32_t ya, int32_t xb, int32_t yb);
 extern void radeon_disable_clipping_mmio(struct fb_info *info);
 
 extern int32_t radeon_setup_for_cpu_to_screen_alpha_texture_mmio(struct fb_info *info,
-                                                                 int op, int red, int green, int blue,
-                                                                 int alpha, int maskFormat, int dstFormat,
-                                                                 uint8_t *alphaPtr, int alphaPitch,
-                                                                 int width, int height, int32_t flags);
+                                                                 int32_t op, int32_t red, int32_t green, int32_t blue,
+                                                                 int32_t alpha, int32_t maskFormat, int32_t dstFormat,
+                                                                 uint8_t *alphaPtr, int32_t alphaPitch,
+                                                                 int32_t width, int32_t height, int32_t flags);
 extern int32_t radeon_setup_for_cpu_to_screen_texture_mmio(struct fb_info *info, int32_t op,
                                                            uint32_t srcFormat, uint32_t dstFormat,
                                                            uint8_t *texPtr, int32_t texPitch,
@@ -647,9 +647,9 @@ extern void radeon_subsequent_cpu_to_screen_texture_mmio(struct fb_info *info,
                                                          int32_t srcx, int32_t srcy,
                                                          int32_t width, int32_t height);
 /* Cursor functions */
-extern void radeon_set_cursor_colors(struct fb_info *info, int bg, int fg);
-extern void radeon_set_cursor_position(struct fb_info *info, int x, int y);
-extern void radeon_load_cursor_image(struct fb_info *info, unsigned short *mask, unsigned short *data, int zoom);
+extern void radeon_set_cursor_colors(struct fb_info *info, int32_t bg, int32_t fg);
+extern void radeon_set_cursor_position(struct fb_info *info, int32_t x, int32_t y);
+extern void radeon_load_cursor_image(struct fb_info *info, unsigned short *mask, unsigned short *data, int32_t zoom);
 extern void radeon_hide_cursor(struct fb_info *info);
 extern void radeon_show_cursor(struct fb_info *info);
 extern long radeon_cursor_init(struct fb_info *info);
@@ -657,8 +657,8 @@ extern long radeon_cursor_init(struct fb_info *info);
 /* Other functions */
 extern int32_t radeon_screen_blank(struct radeonfb_info *rinfo, int32_t blank, int32_t mode_switch);
 extern void radeon_write_mode(struct radeonfb_info *rinfo, struct radeon_regs *mode, int32_t reg_only);
-int radeonfb_setcolreg(unsigned regno, unsigned red, unsigned green,
-                       unsigned blue, unsigned transp, struct fb_info *info);
+int32_t radeonfb_setcolreg(uint32_t regno, uint32_t red, uint32_t green,
+                       uint32_t blue, uint32_t transp, struct fb_info *info);
 extern int32_t radeonfb_pci_register(int32_t handle, const struct pci_device_id *ent);
 extern void radeonfb_pci_unregister(void);
